requirements.txt
# This file was autogenerated by uv via the following command:
#    uv export --no-hashes
affine==2.4.0
aiohappyeyeballs==2.4.6
aiohttp==3.11.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
attrs==25.1.0
babel==2.17.0
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
click-plugins==1.1.1
cligj==0.7.2
colorama==0.4.6 ; platform_system == 'Windows' or sys_platform == 'win32'
covjson-pydantic==0.5.0
dateparser==1.2.1
execnet==2.1.1
filelock==3.17.0
flask==3.1.0
frozenlist==1.5.0
greenlet==3.1.1 ; (python_full_version < '3.14' and platform_machine == 'AMD64') or (python_full_version < '3.14' and platform_machine == 'WIN32') or (python_full_version < '3.14' and platform_machine == 'aarch64') or (python_full_version < '3.14' and platform_machine == 'amd64') or (python_full_version < '3.14' and platform_machine == 'ppc64le') or (python_full_version < '3.14' and platform_machine == 'win32') or (python_full_version < '3.14' and platform_machine == 'x86_64')
h11==0.14.0
idna==3.10
iniconfig==2.0.0
itsdangerous==2.2.0
jinja2==3.1.6
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
lark==1.2.2
markupsafe==3.0.2
multidict==6.1.0
numpy==2.2.3
packaging==24.2
pluggy==1.5.0
propcache==0.3.0
pydantic==2.10.6
pydantic-core==2.27.2
pygeoapi @ git+https://github.com/webb-ben/pygeoapi.git@821f984402131fa352a25a491832feb12d83161c
pygeofilter==0.3.1
pygeoif==1.5.1
pyparsing==3.2.1
pyproj==3.7.1
pytest==8.3.4
pytest-xdist==3.6.1
python-dateutil==2.9.0.post0
pytz==2025.1
pyyaml==6.0.2
rasterio==1.4.3
redis==5.2.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rpds-py==0.23.1
shapely==2.0.7
six==1.17.0
sniffio==1.3.1
sqlalchemy==2.0.39
starlette==0.46.0
tinydb==4.8.2
typing-extensions==4.12.2
tzdata==2025.1 ; platform_system == 'Windows'
tzlocal==5.3.1
urllib3==2.3.0
uvicorn==0.34.0
werkzeug==3.1.3
yarl==1.18.3
# This file was autogenerated by uv via the following command:
#    uv export --no-hashes
affine==2.4.0
aiohappyeyeballs==2.4.6
aiohttp==3.11.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
attrs==25.1.0
babel==2.17.0
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
click-plugins==1.1.1
cligj==0.7.2
colorama==0.4.6 ; platform_system == 'Windows' or sys_platform == 'win32'
covjson-pydantic==0.5.0
dateparser==1.2.1
execnet==2.1.1
filelock==3.17.0
flask==3.1.0
frozenlist==1.5.0
geojson-pydantic==1.2.0
greenlet==3.1.1 ; (python_full_version < '3.14' and platform_machine == 'AMD64') or (python_full_version < '3.14' and platform_machine == 'WIN32') or (python_full_version < '3.14' and platform_machine == 'aarch64') or (python_full_version < '3.14' and platform_machine == 'amd64') or (python_full_version < '3.14' and platform_machine == 'ppc64le') or (python_full_version < '3.14' and platform_machine == 'win32') or (python_full_version < '3.14' and platform_machine == 'x86_64')
h11==0.14.0
idna==3.10
iniconfig==2.0.0
itsdangerous==2.2.0
jinja2==3.1.6
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
lark==1.2.2
markupsafe==3.0.2
multidict==6.1.0
numpy==2.2.3
packaging==24.2
pluggy==1.5.0
propcache==0.3.0
pydantic==2.10.6
pydantic-core==2.27.2
pygeoapi @ git+https://github.com/webb-ben/pygeoapi.git@821f984402131fa352a25a491832feb12d83161c
pygeofilter==0.3.1
pygeoif==1.5.1
pyparsing==3.2.1
pyproj==3.7.1
pytest==8.3.4
pytest-xdist==3.6.1
python-dateutil==2.9.0.post0
pytz==2025.1
pyyaml==6.0.2
rasterio==1.4.3
redis==5.2.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rpds-py==0.23.1
shapely==2.0.7
six==1.17.0
sniffio==1.3.1
sqlalchemy==2.0.39
starlette==0.46.0
tinydb==4.8.2
typing-extensions==4.12.2
tzdata==2025.1 ; platform_system == 'Windows'
tzlocal==5.3.1
urllib3==2.3.0
uvicorn==0.34.0
werkzeug==3.1.3
yarl==1.18.3
# This file was autogenerated by uv via the following command:
#    uv export --no-hashes
affine==2.4.0
aiohappyeyeballs==2.4.6
aiohttp==3.11.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
attrs==25.1.0
babel==2.17.0
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
click-plugins==1.1.1
cligj==0.7.2
colorama==0.4.6 ; platform_system == 'Windows' or sys_platform == 'win32'
covjson-pydantic==0.5.0
dateparser==1.2.1
deprecated==1.2.18
execnet==2.1.1
filelock==3.17.0
flask==3.1.0
frozenlist==1.5.0
geojson-pydantic==1.2.0
googleapis-common-protos==1.56.1
greenlet==3.1.1 ; (python_full_version < '3.14' and platform_machine == 'AMD64') or (python_full_version < '3.14' and platform_machine == 'WIN32') or (python_full_version < '3.14' and platform_machine == 'aarch64') or (python_full_version < '3.14' and platform_machine == 'amd64') or (python_full_version < '3.14' and platform_machine == 'ppc64le') or (python_full_version < '3.14' and platform_machine == 'win32') or (python_full_version < '3.14' and platform_machine == 'x86_64')
grpcio==1.71.0
h11==0.14.0
idna==3.10
importlib-metadata==8.5.0
iniconfig==2.0.0
itsdangerous==2.2.0
jinja2==3.1.6
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
lark==1.2.2
markupsafe==3.0.2
multidict==6.1.0
numpy==2.2.3
opentelemetry-api==1.30.0
opentelemetry-exporter-otlp==1.30.0
opentelemetry-exporter-otlp-proto-common==1.30.0
opentelemetry-exporter-otlp-proto-grpc==1.30.0
opentelemetry-exporter-otlp-proto-http==1.30.0
opentelemetry-instrumentation==0.51b0
opentelemetry-instrumentation-aiohttp-client==0.51b0
opentelemetry-proto==1.30.0
opentelemetry-sdk==1.30.0
opentelemetry-semantic-conventions==0.51b0
opentelemetry-util-http==0.51b0
packaging==24.2
pluggy==1.5.0
propcache==0.3.0
protobuf==5.29.3
pydantic==2.10.6
pydantic-core==2.27.2
pygeoapi @ git+https://github.com/webb-ben/pygeoapi.git@821f984402131fa352a25a491832feb12d83161c
pygeofilter==0.3.1
pygeoif==1.5.1
pyparsing==3.2.1
pyproj==3.7.1
pytest==8.3.4
pytest-xdist==3.6.1
python-dateutil==2.9.0.post0
pytz==2025.1
pyyaml==6.0.2
rasterio==1.4.3
redis==5.2.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rpds-py==0.23.1
shapely==2.0.7
six==1.17.0
sniffio==1.3.1
sqlalchemy==2.0.39
starlette==0.46.0
tinydb==4.8.2
typing-extensions==4.12.2
tzdata==2025.1 ; platform_system == 'Windows'
tzlocal==5.3.1
urllib3==2.3.0
uvicorn==0.34.0
werkzeug==3.1.3
wrapt==1.17.2
yarl==1.18.3
zipp==3.21.0
